+++ geom/type_polygon.go
@@ -49,9 +49,6 @@ func NewPolygon(rings []LineString, opts ...ConstructorOption) (Polygon, error) 
 
 	ctorOpts := newOptionSet(opts)
 	if err := validatePolygon(rings, ctorOpts); err != nil {
-		if ctorOpts.omitInvalid {
-			return Polygon{}.ForceCoordinatesType(ctype), nil
-		}
 		return Polygon{}, err
 	}
 	return Polygon{rings, ctype}, nil